Adding a way to show # of tiles you can move w/o encounter
Changed up the program quite a bit, added a calculator that can tell you exactly how many tiles you can move in the grass before encountering a wild Pokemon, assuming the RNG function is not called extra times, and that you are always walking. I have not added in a way for it to calculate while you are running or while you are on the acro/mach bikes. Once again, this feature only works in Ruby/Sapphire as far as I can tell.